Rename ApplicationMessage to CustomApplicationMessage and update references; enhance NewProcessRequestDTOMapper with additional mapping configurations

+++ src/main/java/com/bpm/mrceprocess/service/impl/ProcessActionServiceImpl.java
@@ -102,6 +102,10 @@ public class ProcessActionServiceImpl implements ProcessActionService {
         GeneralInformationHistory history = historyRepository.findById(historyId)
                 .orElseThrow(() -> new ApplicationException(ApplicationMessage.NOT_FOUND));
 
+        if (!history.isDraft()) {
+            throw new ApplicationException(ApplicationMessage.BAD_REQUEST, "This is not a draft process");
+        }
+
         GeneralInformation generalInformation = history.getGeneralInformation();
         if (generalInformation.getHistories().size() > 1) {
             historyRepository.delete(history);
